requirements.txt
streamlit>=1.33.0
python-dotenv>=1.0.1
langchain>=0.1.20
langchain-openai>=0.1.7
langchain-community>=0.0.38
langchain-ollama>=0.1.0
pypdf>=4.2.0
chromadb>=0.5.0
faiss-cpu>=1.8.0.post1
tiktoken>=0.7.0
